requirements.txt
flask==3.0.0
flask-cors==4.0.0
pytesseract==0.3.10
opencv-python-headless==4.8.1.78
Pillow==10.1.0
numpy==1.24.3
